Added a new section - tiles without backgrounds/borders
Also added a new css rule (.hero-tile) for applying to a section for identifying that the first title of the section will be a 'hero tile' (i.e. span the width of the page at med width). As of now all the rule does it constrain the text of the tile to a specific width at med width.